mcp-script/requirements.txt
# MCP Neo4j Integration Requirements
mcp-neo4j-cypher>=0.3.1
openai>=1.0.0
python-dotenv>=1.0.0
neo4j>=5.0.0
mcp>=1.0.0
httpx>=0.25.0
